🔧 Chore: .cpp, .h, 라이브러리 파일을 각 폴더로 이동
기존에 같이 있던 모든 파일을 .cpp, .h, 라이브러리에 따라 폴더로 구분하여 정리

+++ source/cpp/Course.cpp
@@ -7,7 +7,7 @@
 namespace Course 
 {
 
-  std::string convertToGrade(const int& grade)
+  std::string convertToGrade(const double& grade)
   {
     std::string g[10] = {"A+", "A", "B+", "B", "C+", "C", "D+", "D", "F", "P"};
     int index = 0;
